refactor(style): move style to core crate
