+++ crates/core/src/style/mod.rs
@@ -18,3 +18,17 @@ impl ToString for Style {
         .into()
     }
 }
+
+mod basic;
+mod column;
+mod flex;
+mod grid;
+mod widget;
+
+pub use self::{
+    basic::{ContainerStyle, SizedBoxStyle},
+    column::MultiColumnStyle,
+    flex::{AlignStyle, FlexStyle},
+    grid::GridStyle,
+    widget::{ImageSrc, TextStyle},
+};
