+++ crates/epm/src/template.rs
@@ -18,5 +18,5 @@ path = "/var/folders/5_/m9p8c1mn6k34_x77p79v15z00000gn/T/hello-world/src"
 
 [[children.children]]
 path = "/var/folders/5_/m9p8c1mn6k34_x77p79v15z00000gn/T/hello-world/src/lib.rs"
-content = [117, 115, 101, 32, 101, 108, 118, 105, 115, 58, 58, 123, 10, 32, 32, 32, 32, 100, 114, 105, 118, 101, 114, 58, 58, 123, 68, 114, 105, 118, 101, 114, 44, 32, 87, 101, 98, 68, 114, 105, 118, 101, 114, 125, 44, 10, 32, 32, 32, 32, 103, 101, 115, 116, 117, 114, 101, 58, 58, 123, 71, 101, 115, 116, 117, 114, 101, 44, 32, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 125, 44, 10, 32, 32, 32, 32, 112, 114, 101, 108, 117, 100, 101, 58, 58, 42, 44, 10, 32, 32, 32, 32, 119, 105, 100, 103, 101, 116, 115, 58, 58, 123, 108, 97, 121, 111, 117, 116, 115, 58, 58, 67, 101, 110, 116, 101, 114, 44, 32, 84, 101, 120, 116, 44, 32, 84, 101, 120, 116, 83, 116, 121, 108, 101, 125, 44, 10, 125, 59, 10, 10, 35, 91, 112, 97, 103, 101, 93, 10, 115, 116, 114, 117, 99, 116, 32, 73, 110, 100, 101, 120, 59, 10, 10, 105, 109, 112, 108, 32, 76, 105, 102, 101, 67, 121, 99, 108, 101, 60, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 60, 67, 101, 110, 116, 101, 114, 62, 62, 32, 102, 111, 114, 32, 73, 110, 100, 101, 120, 32, 123, 10, 32, 32, 32, 32, 102, 110, 32, 99, 114, 101, 97, 116, 101, 40, 38, 115, 101, 108, 102, 41, 32, 45, 62, 32, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 60, 67, 101, 110, 116, 101, 114, 62, 32, 123, 10, 32, 32, 32, 32, 32, 32, 32, 32, 108, 101, 116, 32, 99, 101, 110, 116, 101, 114, 32, 61, 32, 67, 101, 110, 116, 101, 114, 32, 123, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 99, 104, 105, 108, 100, 58, 32, 84, 101, 120, 116, 32, 123, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 116, 101, 120, 116, 58, 32, 34, 80, 105, 110, 107, 32, 105, 115, 32, 116, 104, 101, 32, 80, 105, 103, 33, 33, 33, 34, 46, 105, 110, 116, 111, 40, 41, 44, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 115, 116, 121, 108, 101, 58, 32, 84, 101, 120, 116, 83, 116, 121, 108, 101, 58, 58, 100, 101, 102, 97, 117, 108, 116, 40, 41, 44, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 125, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 46, 105, 110, 116, 111, 40, 41, 44, 10, 32, 32, 32, 32, 32, 32, 32, 32, 125, 59, 10, 10, 32, 32, 32, 32, 32, 32, 32, 32, 108, 101, 116, 32, 109, 117, 116, 32, 100, 101, 116, 101, 99, 116, 111, 114, 32, 61, 32, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 58, 58, 110, 101, 119, 40, 99, 101, 110, 116, 101, 114, 41, 59, 10, 32, 32, 32, 32, 32, 32, 32, 32, 100, 101, 116, 101, 99, 116, 111, 114, 46, 114, 101, 103, 105, 115, 116, 101, 114, 40, 71, 101, 115, 116, 117, 114, 101, 58, 58, 84, 97, 112, 44, 32, 124, 95, 124, 32, 87, 101, 98, 68, 114, 105, 118, 101, 114, 58, 58, 97, 108, 101, 114, 116, 40, 34, 72, 101, 108, 108, 111, 32, 102, 114, 111, 109, 32, 114, 117, 115, 116, 33, 34, 41, 41, 59, 10, 32, 32, 32, 32, 32, 32, 32, 32, 100, 101, 116, 101, 99, 116, 111, 114, 10, 32, 32, 32, 32, 125, 10, 125, 10]
+content = [117, 115, 101, 32, 101, 108, 118, 105, 115, 58, 58, 123, 10, 32, 32, 32, 32, 100, 114, 105, 118, 101, 114, 58, 58, 123, 68, 114, 105, 118, 101, 114, 44, 32, 87, 101, 98, 68, 114, 105, 118, 101, 114, 125, 44, 10, 32, 32, 32, 32, 103, 101, 115, 116, 117, 114, 101, 58, 58, 123, 71, 101, 115, 116, 117, 114, 101, 44, 32, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 125, 44, 10, 32, 32, 32, 32, 112, 114, 101, 108, 117, 100, 101, 58, 58, 42, 44, 10, 32, 32, 32, 32, 115, 116, 121, 108, 101, 58, 58, 84, 101, 120, 116, 83, 116, 121, 108, 101, 44, 10, 32, 32, 32, 32, 119, 105, 100, 103, 101, 116, 115, 58, 58, 123, 108, 97, 121, 111, 117, 116, 115, 58, 58, 67, 101, 110, 116, 101, 114, 44, 32, 84, 101, 120, 116, 125, 44, 10, 125, 59, 10, 10, 35, 91, 112, 97, 103, 101, 93, 10, 115, 116, 114, 117, 99, 116, 32, 73, 110, 100, 101, 120, 59, 10, 10, 105, 109, 112, 108, 32, 76, 105, 102, 101, 67, 121, 99, 108, 101, 60, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 60, 67, 101, 110, 116, 101, 114, 62, 62, 32, 102, 111, 114, 32, 73, 110, 100, 101, 120, 32, 123, 10, 32, 32, 32, 32, 102, 110, 32, 99, 114, 101, 97, 116, 101, 40, 38, 115, 101, 108, 102, 41, 32, 45, 62, 32, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 60, 67, 101, 110, 116, 101, 114, 62, 32, 123, 10, 32, 32, 32, 32, 32, 32, 32, 32, 108, 101, 116, 32, 99, 101, 110, 116, 101, 114, 32, 61, 32, 67, 101, 110, 116, 101, 114, 32, 123, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 99, 104, 105, 108, 100, 58, 32, 84, 101, 120, 116, 32, 123, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 116, 101, 120, 116, 58, 32, 34, 80, 105, 110, 107, 32, 105, 115, 32, 116, 104, 101, 32, 80, 105, 103, 33, 33, 33, 34, 46, 105, 110, 116, 111, 40, 41, 44, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 115, 116, 121, 108, 101, 58, 32, 84, 101, 120, 116, 83, 116, 121, 108, 101, 58, 58, 100, 101, 102, 97, 117, 108, 116, 40, 41, 44, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 125, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 46, 105, 110, 116, 111, 40, 41, 44, 10, 32, 32, 32, 32, 32, 32, 32, 32, 125, 59, 10, 10, 32, 32, 32, 32, 32, 32, 32, 32, 108, 101, 116, 32, 109, 117, 116, 32, 100, 101, 116, 101, 99, 116, 111, 114, 32, 61, 32, 71, 101, 115, 116, 117, 114, 101, 68, 101, 116, 101, 99, 116, 111, 114, 58, 58, 110, 101, 119, 40, 99, 101, 110, 116, 101, 114, 41, 59, 10, 32, 32, 32, 32, 32, 32, 32, 32, 100, 101, 116, 101, 99, 116, 111, 114, 46, 114, 101, 103, 105, 115, 116, 101, 114, 40, 71, 101, 115, 116, 117, 114, 101, 58, 58, 84, 97, 112, 44, 32, 124, 95, 124, 32, 87, 101, 98, 68, 114, 105, 118, 101, 114, 58, 58, 97, 108, 101, 114, 116, 40, 34, 72, 101, 108, 108, 111, 32, 102, 114, 111, 109, 32, 114, 117, 115, 116, 33, 34, 41, 41, 59, 10, 32, 32, 32, 32, 32, 32, 32, 32, 100, 101, 116, 101, 99, 116, 111, 114, 10, 32, 32, 32, 32, 125, 10, 125, 10]
 "#;
