+++ src/widgets/mod.rs
@@ -3,7 +3,4 @@ mod image;
 pub mod layouts;
 mod text;
 
-pub use {
-    image::{Image, ImageSrc},
-    text::{Text, TextStyle},
-};
+pub use {image::Image, text::Text};
